refactor: apply eslint
