+++ .eslintrc.json
@@ -1,0 +1,17 @@
+{
+  "extends": ["airbnb", "prettier", "plugin:node/recommended"],
+  "plugins": ["prettier"],
+  "rules": {
+    "prettier/prettier": "off",
+    "no-unused-vars": "error",
+    "no-console": "warn",
+    "func-names": "off",
+    "no-process-exit": "off",
+    "object-shorthand": "off",
+    "class-methods-use-this": "off",
+    "node/no-unsupported-features/es-syntax": "off"
+  },
+  "parserOptions": {
+    "sourceType": "module"
+  }
+}
